backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg[binary]==3.2.9
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
stripe==6.7.0
sendgrid==6.10.0
python-dotenv==1.0.0
requests==2.31.0
